pyproject.toml
[project]
name = "a2a-rag-agent"
version = "0.1.0"
description = "A2A RAG Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "langchain-openai>=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "a2a-sdk>=0.3.0",
    "langchain-ollama>=0.3.8",
    "pydantic-settings>=2.10.1",
    "langchain>=0.3.27",
    "langchain-text-splitters>=0.3.11",
    "numpy>=2.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['src/a2a_rag_agent']
